Gagnant done

+++ src/sti_fork_lld_lldi_lfork.c
@@ -23,8 +23,6 @@ void  sti(t_process *processor, unsigned char *map, int iz, t_player *pl, t_vizD
 	convert(&arg1);
 	n = 0;
 	arg2 = ft_strnew(8);
-
-
 	if (ft_strlen(processor->arg2) == 2)
 	{
 		if (ft_strlen(processor->arg3) == 2)
@@ -40,14 +38,13 @@ void  sti(t_process *processor, unsigned char *map, int iz, t_player *pl, t_vizD
 	{
 		if (ft_strlen(processor->arg3) == 4)
 		{
-			i = ((short)(((short)ft_atoi_base(processor->arg2, 16) + (short)(ft_atoi_base(processor->arg3, 16))))% IDX_MOD) * 2 + processor->cur_pos;
+			i = ((((short)ft_atoi_base(processor->arg2, 16) + (short)(ft_atoi_base(processor->arg3, 16))))% IDX_MOD) * 2 + processor->cur_pos;
 		}
 		else
 		{
 			i = ((short)(((short)ft_atoi_base(processor->arg2, 16) + processor->reg[ft_atoi_base(processor->arg3, 16) - 1] ))% IDX_MOD) * 2 + processor->cur_pos;
 		}
 	}
-
 	if (i < 0)
 		i = MEM_SIZE * 2 + i;
 	while (n < 8)
